Understanding of m-trees improved and thus so flaws in the implementation arise.
The leaf nodes together contain every data entry so they now have an std::array containing shared pointers while the objects used to parent higher nodes use weak_ptrs. Additionally the capacity is now passed into the template.

Also added constructors to the tree node based structs. If you construct a tree node with a distance to parent it initialises the variant to use a routing object else it initialises it to be a leaf. This will probably be changed later but for now it's sufficient