Implemented first of the split functions again.
Also not sure about this function map, as I need to std::bind the implicit this pointer to use the functions. However it should be less verbose than implementing a "split object". But it does come with a bigger memory footprint, unless there is a lazy initialisation....

Either way a demonstration of adding a function to the mapping is shown.