Closer to knn working
knn_query can now prioritise between invalid ids and valid ids of the same distance. Removal logic however doesn't. Also using the default constructed ID as the invalid one is an ugly hack.

Boost::optional seems overkill for this, but it does make the most sense from a design perspective. So coming next should be a version with optional added